requirements.txt
langgraph>=0.0.55
langchain>=0.1.17
langchain-openai>=0.1.7
langchain-anthropic>=0.1.11
langchain-core>=0.1.52
tiktoken>=0.6.0
python-dotenv>=1.0.0
pydantic>=2.5.0
click>=8.1.7
PyYAML>=6.0.1
